+++ include/UKPmanager.hpp
@@ -3,12 +3,14 @@
 #include "button.hpp"
 #include "options.hpp"
 #include "binary_output.hpp"
+#include "_7segment.hpp"
 
 class UKPmanager
 {
 private:
     button btn[BUTTONS];
     binary_output out;
+    _7segment seg;
 //if profiles are enabled track when they change
 #if PROFILES
     int last_profile_state;